refactor: Score, Side 패키지 game으로 이동
