+++ src/main/java/chess/domain/board/Board.java
@@ -1,6 +1,6 @@
 package chess.domain.board;
 
-import chess.domain.Side;
+import chess.domain.game.Side;
 import chess.domain.piece.Blank;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
