+++ src/test/java/chess/domain/game/SideTest.java
@@ -1,8 +1,9 @@
-package chess.domain;
+package chess.domain.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,9 +12,10 @@ class SideTest {
     @Test
     @DisplayName("턴 교체")
     void changeTurn() {
-        assertThat(Side.WHITE.changeTurn()).isEqualTo(Side.BLACK);
+        Assertions.assertThat(Side.WHITE.changeTurn()).isEqualTo(Side.BLACK);
         assertThat(Side.BLACK.changeTurn()).isEqualTo(Side.WHITE);
-        assertThatThrownBy(() -> Side.NONE.changeTurn())
-                .isInstanceOf(UnsupportedOperationException.class);
+        assertThatThrownBy(() -> {
+            Side side = Side.NONE.changeTurn();
+        }).isInstanceOf(UnsupportedOperationException.class);
     }
 }
