+++ src/test/java/chess/domain/gamestate/GameSetTest.java
@@ -2,9 +2,9 @@ package chess.domain.gamestate;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.Side;
 import chess.domain.board.Board;
 import chess.domain.board.BoardInitializer;
+import chess.domain.game.Side;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
